Create questions and answers. Answer list shown in the view is fake.

+++ src/main/java/may/bestpoll/service/SequenceService.java
@@ -9,5 +9,5 @@ public interface SequenceService
 	 * Returns a named counter after incrementing it.
 	 * It will be created (with initial value of 1) if it doesn't exist.
 	 */
-	long getCounter(String name);
+	int getCounter(String name);
 }
